broke out rain-detail
