+++ app/app.module.ts
@@ -1,11 +1,19 @@
 import { NgModule }      from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule }   from '@angular/forms';
 
 import { AppComponent } from './app.component';
+import { RainDetailComponent } from './rain-detail.component';
 
 @NgModule({
-  imports:      [ BrowserModule ],
-  declarations: [ AppComponent ],
+  imports:      [
+    BrowserModule,
+    FormsModule
+  ],
+  declarations: [
+    AppComponent,
+    RainDetailComponent
+  ],
   bootstrap:    [ AppComponent ]
 })
 export class AppModule { }
