+++ app/rain.ts
@@ -1,0 +1,7 @@
+export class Rain {
+  recognize: string;
+  inquiry: string;
+  investigate: string;
+  nonid: boolean;
+}
+
